Add tests for LVM_VolGroup.annotations
